Far stronger work on UpValues Not working quite yet but we'll get there.
Bit too much in this one...

+++ crates/runtime/src/table.rs
@@ -22,7 +22,7 @@ impl TabValueExt for Value {
 	}
 
 	fn table(table: Table) -> Self {
-		let idx = table.get_ptr() as usize;
+		let idx = table.get_ptr() as u32;
 		let mut val = unsafe { Value::idx(idx) };
 		val.set_tag(ValueTag::Table);
 		val
